pyproject.toml
[tool.poetry]
name = "weirdo_collector"
version = "0.1.0"
description = "Weirdos, collect them all"
authors = ["Weirdo Watcher <weirdowatcher420@gmail.com>"]

[tool.poetry.scripts]
weirdo-collector = 'weirdo_collector.cli:main'

[tool.poetry.dependencies]
python = "^3.8"
twint = "^2.1.20"
pyyaml = "^5.3.1"
click = "^7.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
